Enable warnings and fix them, all of them
This involved adding prototypes for most functions and fixing
all errors revealed by this. The outcome should be a much more
stable Micropolis, as now all functions are called with their
defined arguments and also have a defined return value.
This should fix debian bug #522529 in a different way than the
patch in the bug description